+++ modules/effect-cards/effect-cards.js
@@ -81,11 +81,7 @@ export default function EffectCards({
 
 			const scaleString =
 				progress < 0 ? `${1 + (1 - scale) * progress}` : `${1 - (1 - scale) * progress}`;
-			const transform = `
-        translate3d(${tX}, ${tY}, ${tZ}px)
-        rotateZ(${rotate}deg)
-        scale(${scaleString})
-      `;
+			const transform = `translate3d(${tX}, ${tY}, ${tZ}px) rotateZ(${rotate}deg) scale(${scaleString})`;
 
 			// if (params.slideShadows) {
 			//   // Set shadows